refactor(tools): separate gzip functions from tools.utils
