+++ kong-3.6.0-0.rockspec
@@ -164,6 +164,7 @@ build = {
     ["kong.tools.protobuf"] = "kong/tools/protobuf.lua",
     ["kong.tools.mime_type"] = "kong/tools/mime_type.lua",
     ["kong.tools.request_aware_table"] = "kong/tools/request_aware_table.lua",
+    ["kong.tools.gzip"] = "kong/tools/gzip.lua",
     ["kong.tools.string"] = "kong/tools/string.lua",
     ["kong.tools.table"] = "kong/tools/table.lua",
     ["kong.tools.yield"] = "kong/tools/yield.lua",
